myproject/requirements.txt
asgiref==3.3.4
Django==3.2.2
django-filter==2.4.0
django-tables2==2.3.4
django-widget-tweaks==1.4.8
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
